refactor: 닉네임 파일 common으로 경로변경
